Move darkModeStyle to new file
